fix: chat_id в BigInteger (Telegram ID > 2.1 млрд не влезал в Integer)

+++ backend/app/models/telegram.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, DateTime, Boolean
+from sqlalchemy import Column, Integer, BigInteger, String, DateTime, Boolean
 from datetime import datetime
 from .vehicles import Base
 
@@ -7,7 +7,7 @@ class TelegramLink(Base):
 
     id = Column(Integer, primary_key=True, autoincrement=True)
     user_id = Column(Integer, nullable=False, index=True)
-    chat_id = Column(Integer, nullable=True)
+    chat_id = Column(BigInteger, nullable=True)
     code = Column(String, nullable=True)
     code_expires_at = Column(DateTime, nullable=True)
     telegram_username = Column(String, nullable=True)
